docs for precond
